SARdespeckling/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pyhd8ed1ab_0
astor=0.8.1=pyh9f0ad1d_0
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pypi_0
boost-cpp=1.72.0=h89d28cc_2
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.5.30=h5b45459_0
certifi=2021.5.30=py36ha15d459_0
cfitsio=3.470=h0af3d06_7
cftime=1.0.4.2=pypi_0
click=7.1.2=pyh9f0ad1d_0
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_0
colorama=0.4.4=pypi_0
curl=7.68.0=h4496350_0
cycler=0.10.0=py_2
dask=2021.3.0=pypi_0
data-complexity=0.1.3=pypi_0
dataclasses=0.8=pyh787bdff_0
decorator=4.4.2=pypi_0
equi7grid=0.1.0=pypi_0
expat=2.4.1=h39d44d4_0
fiona=1.8.13=py36hb7fdc2d_0
freetype=2.10.4=h546665d_1
freexl=1.0.6=ha8e266a_0
gast=0.5.0=pyhd8ed1ab_0
gdal=3.0.4=py36h2fee047_0
geographiclib=1.52=pypi_0
geopandas=0.6.3=py_0
geos=3.8.0=he025d50_0
geotiff=1.5.1=h9633a31_8
gettext=0.19.8.1=h1a89ca6_1005
glib=2.68.3=h0e60522_0
glib-tools=2.68.3=h0e60522_0
google-pasta=0.2.0=pyh8c360ce_0
gower=0.0.5=pypi_0
grpcio=1.38.1=py36h4374274_0
h5py=2.10.0=nompi_py36h422b98e_102
hdf4=4.2.15=h0e5069d_3
hdf5=1.10.5=nompi_h4d922ce_1114
icu=64.2=he025d50_1
importlib-metadata=4.6.3=py36ha15d459_0
intel-openmp=2021.3.0=h57928b3_3372
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.0=pypi_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
kealib=1.4.10=hf7dc31f_1005
keras=2.3.1=py36h21ff451_0
keras-applications=1.0.8=py_1
keras-metrics=1.1.0=pypi_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py36he95197e_1
krb5=1.16.4=hdd46e55_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libcurl=7.68.0=h4496350_0
libffi=3.3=h0e60522_2
libgdal=3.0.4=h90de84b_0
libglib=2.68.3=h1e62bf3_0
libgpuarray=0.7.6=h8ffe710_1003
libiconv=1.16=he774522_0
libkml=1.3.0=h9d95caa_1012
liblapack=3.9.0=8_mkl
libnetcdf=4.7.3=nompi_hc957ea6_101
libpng=1.6.37=h1d00b33_2
libpq=12.1=hb0bdaea_0
libprotobuf=3.17.2=h7755175_1
libspatialindex=1.9.3=h39d44d4_4
libspatialite=4.3.0a=hd027ac7_1034
libssh2=1.8.2=h642c060_2
libtiff=4.1.0=h21b02b4_3
libwebp=1.0.2=hfa6e2cd_5
libxml2=2.9.12=hf5bbc77_0
lz4-c=1.8.3=he025d50_1001
m2w64-expat=2.1.1=2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gettext=0.19.7=2
m2w64-gmp=6.1.0=2
m2w64-libiconv=1.14=6
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-xz=5.2.2=2
mako=1.1.4=pyh44b312d_0
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=2.0.1=py36h68aa20f_0
matplotlib=3.1.2=py36_1
matplotlib-base=3.1.2=py36h2981e6d_1
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
munch=2.5.0=py_0
netcdf4=1.5.3=pypi_0
numpy=1.19.5=py36h4b40d73_2
opencv-python=4.5.3.56=pypi_0
openjpeg=2.3.1=h48faf41_3
openssl=1.1.1k=h8ffe710_0
pandas=0.25.3=py36he350917_0
parso=0.8.1=pypi_0
pcre=8.45=h0e60522_0
pickleshare=0.7.5=pypi_0
pillow=8.3.2=pypi_0
pip=21.2.3=pyhd8ed1ab_0
poppler=0.67.0=h1707e21_8
poppler-data=0.4.10=0
postgresql=12.1=h06f7779_0
proj=6.3.0=ha7a8c7b_0
prompt-toolkit=3.0.11=pypi_0
protobuf=3.17.2=py36he2d232f_0
pygments=2.7.4=pypi_0
pygpu=0.7.6=py36h6434af4_1002
pyparsing=2.4.7=pyh9f0ad1d_0
pyproj=2.4.2.post1=py36hc1560cf_1
pyqt=5.9.2=py36h6538335_4
pyreadline=2.1=py36ha15d459_1003
python=3.6.13=h39d44d4_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.6=2_cp36m
pytileproj=0.0.16=pypi_0
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.4.1=py36h68aa20f_0
qt=5.9.7=h506e8af_3
rtree=0.9.3=py36hbf79ddb_0
scikit-learn=0.22=py36h7208079_1
scipy=1.5.3=py36h7ff6e69_0
setuptools=49.6.0=py36ha15d459_3
shapely=1.7.0=py36h2130f3d_0
sip=4.19.8=py36h6538335_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=h8ffe710_0
tbb=2018.0.5=he980bc4_0
tensorboard=1.14.0=py36_0
tensorflow=1.14.0=h39a9291_0
tensorflow-base=1.14.0=py36h9f0ad1d_0
tensorflow-estimator=1.14.0=py36h5ca1d4c_0
termcolor=1.1.0=py_2
theano=1.0.5=py36he2d232f_1
tifffile=2020.9.3=pypi_0
tiledb=1.7.0=hffbbd95_2
tk=8.6.10=h8ffe710_1
toolz=0.11.1=pypi_0
tornado=6.1=py36h68aa20f_1
traitlets=4.3.3=pypi_0
typing_extensions=3.10.0.0=pyha770c72_0
ucrt=10.0.20348.0=h57928b3_0
vc=14.2=hb210afc_5
veranda=0.1.1=pypi_0
vs2015_runtime=14.29.30037=h902a5da_5
vs2017_win-64=19.16.27038=h2e3bad8_2
vswhere=2.8.4=h57928b3_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_0
wincertstore=0.2=py36ha15d459_1006
wrapt=1.12.1=py36h68aa20f_3
xarray=0.15.0=py_0
xerces-c=3.2.2=h6538335_1004
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.4=hd8a0e53_1
